old-tests/submission/Opera/preload/none/preload-none-event-order-poster-delayed.html

<!doctype html>
<title>HTMLVideoElement.preload:none should delay window load for the poster to load</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<p id="log">FAIL (script didn't run)</p>
<video preload="none" poster="../resources/range-request.php?rate=5000&fileloc=poster.png"></video>
<script>
var t = async_test("HTMLVideoElement.preload:none should delay window load for the poster to load", {timeout:3000});
t.step(function() {
var v = document.getElementsByTagName('video')[0];
window.onload = function() {
    t.step(function() {
	assert_equals(getComputedStyle(v,'').width, '102px', 'The video poster should load before window.onload is fired')
    });
    t.done();
}
});
</script>
